Remove DevOS branding and replace with "nixed"
I think that since digga was merged into the DevOS repository,
it's weird to keep using the template's branding. So I replaced
it with the name of this configuration flake, "nixed".

I also cleaned up a bit; the deadnix overlay is no longer needed,
and the bud template and trusted JSON keys are no longer needed
either.