feat: use nvim-treesitter for grammar sources
The nvim-treesitter plugin in nixpkgs uses the nvim-treesitter lockfile,
so all queries are actually synced with the parsers all the time. It
was quite hard to use these generated parsers in such a way where it
could be extended, however, and I wanted to use it without using the
`plugins` option in HM. So here we are.

- Stop fetching latest nvim-treesitter with nvfetcher
- Use lib.extend instead of directly merging lib with other attrsets
- Remove overrides for grammars (except for norg ones, the parser
  must be synced with the plugin)