revert: ditch wezterm, keep using kitty
WezTerm has had a few problems, so I'm gonna stick with Kitty for now.
Stuff like taking exponentially longer to render large windows on
macOS, a little bit of jank when moving around compared to Kitty, and
other small stuff. I like it, but it's a little too weird for me to use
right now; I might switch back in the future but for now I'll stick to
Kitty.