fix: resolve macOS breakage
Haven't used macOS in a while, and have gone back to it. A few things
broke in my config, and I fixed a few and disabled the vscode-lldb
extension because it was failing to build with a cryptic linker
error I do not have the time to resolve right now.