fix(inet): use systemd-resolved for dns resolution
DNS resolution would straight-up not work on some networks, presumably
because they wanted to set the namesevrer in resolv.conf directly but
had no way of doing so, since it is immutable.

This uses systemd-resolved for DNS, which should (hopefully) configure
this automatically so I don't have to keep on doing `sudo -e` to
resolv.conf every time I want to connect to the internet quickly at
SFSU or some random coffee shop.