feat(nvim): replace nil_ls with nixd
For right now, this isn't configured well specifically for this. BUT,
it is a lot faster and it feels a lot nicer already.

I'll get a .nixd.json working once I figure out how the fuck it actually
works, but I think I'm just kind of brain damaged at the moment of
writing.