feat(nvim): replace noice with fidget, bmessages.nvim
I want to start using Neovim in a more vanilla way, and that requires
getting used to cmdline behavior, which is the only reason I used
noice in the first place. So I removed it and updated fidget.nvim
so that notifications would be there.

Also, I brought in bmessages.nvim so that I can view printed changes in
a buffer.